// web/src/__generated__/schemas/CampaignOfferEntitlementsResponse.ts
/**
 * ⚠️ THIS FILE IS AUTO-GENERATED. DO NOT EDIT MANUALLY.
 * All manual edits will be lost when this file is regenerated.
 */

import type { CampaignOfferEntitlement } from './CampaignOfferEntitlement';

export type CampaignOfferEntitlementsResponse = { 
  entitlements?: CampaignOfferEntitlement[]; 
  playerId?: string; 
};